feat: add testes de divisao e multiplicacao e trocar tipo de int pra double

+++ src/FontesCalculadora/CalculadoraAritmetica.java
@@ -2,19 +2,19 @@ package FontesCalculadora;
 
 public class CalculadoraAritmetica {
 	
-	public int somar(int a, int b) {
+	public double somar(double a, double b) {
 		return a + b;
 	}
 	
-	public int subtrair(int a, int b) {
+	public double subtrair(double a, double b) {
 		return a - b;
 	}
 	
-	public int multiplicar(int a, int b) {
+	public double multiplicar(double a, double b) {
 		return a * b;
 	}
 	
-	public int dividir(int a, int b) {
+	public double dividir(double a, double b) {
 		if (b != 0) {
 			return a / b;			
 		} else {
